Macine learing (bird deaseser)/requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0

# HuggingFace (for ConvNeXt, CvT)
transformers>=4.30.0

# timm (for ResNeSt)
timm>=0.9.0

# Data & Metrics
numpy>=1.24.0
scikit-learn>=1.3.0
Pillow>=9.5.0

# Visualization
matplotlib>=3.7.0

# Progress bars
tqdm>=4.65.0
